Remove temporary slurm resume file at the end of resume_program
Prior this commit, the code only had `trap "rm -f ${SLURM_RESUME_FILE_TMP}" EXIT`, which works fine for the script until the `sudo -u` happens. When using `sudo -u` to run a command, it creates a new process with a different user context. The trap that was set in the parent shell does not carry over to this new process. Therefore, when the script ends through the sudo command, the EXIT trap in the original shell never gets executed.

Therefore, this commit add another removal of temporary file at the end of the script

Signed-off-by: Hanwen <hanwenli@amazon.com>